% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DEFAULT_always_positive_targets}
\alias{DEFAULT_always_positive_targets}
\title{Default Always-Positive Targets}
\usage{
DEFAULT_always_positive_targets
}
\description{
Character vector of targets that are expected to always produce a detectable result
(used for internal quality controls).
}
