% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunction.r
\name{most_frequent}
\alias{most_frequent}
\title{Most Frequent Element}
\usage{
most_frequent(data)
}
\arguments{
\item{data}{A vector of elements (numbers or words).}
}
\value{
The most frequent element.
}
\description{
This function finds the most frequent element in a vector.
}
\examples{
most_frequent(c(1, 2, 2, 3, 3, 3, 4))
}
