% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunction.r
\name{prime_factors}
\alias{prime_factors}
\title{Prime Factorization}
\usage{
prime_factors(n)
}
\arguments{
\item{n}{An integer to factorize.}
}
\value{
A vector of prime factors.
}
\description{
This function performs prime factorization of an integer.
}
\examples{
prime_factors(98)
}
