% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunction.r
\name{simple_plot}
\alias{simple_plot}
\title{Simple Data Plot}
\usage{
simple_plot(data, plot_type = "histogram", ...)
}
\arguments{
\item{data}{A numeric vector or data frame to plot.}

\item{plot_type}{The type of plot to create ("histogram", "boxplot", "scatter").}

\item{...}{Additional arguments passed to the plotting functions.}
}
\value{
A plot of the data.
}
\description{
This function creates a simple plot of the given data.
}
\examples{
simple_plot(c(1, 2, 2, 3, 3, 3, 4), "histogram")
}
