% File src/library/tools/man/tools-defunct.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{tools-defunct}
% NOTE: ../R/tools-defunct.R must be synchronized with this!
\alias{tools-defunct}
\title{Defunct Objects in Package tools}
\description{
  The functions or variables listed here are defunct, and may get
  removed as soon as of the next release.
}
% \usage{
% }
\seealso{
  \code{\link{tools-deprecated}},
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{internal}
