\name{tools-defunct}
% NOTE: ../R/tools-defunct.R must be synchronized with this!
\alias{tools-defunct}
\title{Defunct Objects in Package tools}
\description{
  The functions or variables listed here are defunct, and may get
  removed as soon as of the next release.
}
\usage{
}
\seealso{
  \code{\link{tools-deprecated}},
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{internal}
