% File src/library/tools/man/tools-defunct.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2009 R Core Team
% Distributed under GPL 2 or later

\name{tools-defunct}
% NOTE: ../R/tools-defunct.R must be synchronized with this!
\alias{tools-defunct}
\alias{Rd_parse}
\title{Defunct Objects in Package \pkg{tools}}
\description{
  The functions or variables listed here are no longer part of \R as
  they are not needed (any more).
}
\usage{
# Defunct in R 2.10.0
Rd_parse(file, text = NULL)
}
\seealso{
  \code{\link{tools-deprecated}},
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{internal}
