% File src/library/tools/man/tools-deprecated.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{tools-deprecated}
\alias{tools-deprecated}
\title{Deprecated Objects in Package \pkg{tools}}
\description{
  The functions or variables listed here are provided for compatibility
  with older versions of \R only, and may be defunct as soon as of the
  next release.
}
% \usage{
% }
\seealso{
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{misc}
