% File src/library/tools/man/tools-deprecated.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2015 R Core Team
% Distributed under GPL 2 or later

\name{tools-deprecated}
\title{Deprecated Objects in Package \pkg{tools}}
%------ NOTE:  ../R/tools-deprecated.R   must be synchronized with this!
\alias{tools-deprecated}
%------ PLEASE: one \alias{.} for EACH ! (+ one \usage{} & \arguments{} for all)
%% \alias{package.dependencies}
\description{
  The functions or variables listed here are provided for compatibility
  with older versions of \R only, and may be defunct as soon as of the
  next release.
}
%% \usage{%% maybe use  funnname(\dots)   and have only \arguments{ \item{\dots}{} }
%% package.dependencies(x, check = FALSE,
%%                      depLevel =
%%                      c("Depends", "Imports", "Suggests"))
%% }
\seealso{
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{misc}
