\name{tools-deprecated}
\alias{tools-deprecated}
\title{Deprecated Objects in Package tools}
\description{
  The functions or variables listed here are provided for compatibility
  with older versions of \R only, and may be defunct as soon as of the
  next release.
}
\usage{
}
\seealso{
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{misc}
