\name{tools-internal}
\title{Internal tools functions}
\alias{listFilesWithExts}
\alias{sQuote}
\alias{SweaveParseOptions}
\alias{RweaveLatexRuncode}
\alias{RweaveLatexWritedoc}
\alias{RweaveLatexFinish}
\alias{RweaveLatexOptions}
\alias{RweaveChunkPrefix}
\alias{RtangleRuncode}
\alias{RtangleFinish}
\description{
  Internal tools functions.
}
\usage{
listFilesWithExts(dir, exts, path = TRUE)
sQuote(s)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
