% File src/library/tools/man/tools-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{tools-package}
\alias{tools-package}
\alias{tools}
\docType{package}
\title{Tools for Package Development}
\description{
  Tools for package development, administration and documentation.
}
\details{
  This package contains tools for manipulating R packages and their
  documentation.

  For a complete list of functions, use \code{library(help = "tools")}.

}
\author{
  Kurt Hornik and Friedrich Leisch

  Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ programming }
\keyword{ documentation }
